Updated: codegen not to generate loader when there is no project fieldtype

+++ codegen/contenttypes/gen.sh
@@ -9,4 +9,6 @@ dir=`dirname $0`
 path=$1
 sh $dir/fieldtype_loader.sh $path
 go run $dir/gen.go src/$path
-rm $dir/temp/project.go
+if [ -f "$dir/temp/project.go" ]; then
+  rm $dir/temp/project.go
+fi
